JSDK-2954 Requested changes
