--- tsdef/EncodingParameters.d.ts
@@ -1,4 +1,0 @@
-export interface EncodingParameters {
-  maxAudioBitrate?: number | null;
-  maxVideoBitrate?: number | null;
-}
