+++ tsdef/LocalParticipant.d.ts
@@ -1,8 +1,6 @@
-import { NetworkQualityConfiguration, NetworkQualityLevel, NetworkQualityStats } from './NetworkQuality';
-import { EncodingParameters } from './EncodingParameters';
+import { EncodingParameters, LocalTrack, NetworkQualityConfiguration } from './types';
 import { LocalAudioTrackPublication } from './LocalAudioTrackPublication';
 import { LocalDataTrackPublication } from './LocalDataTrackPublication';
-import { LocalTrack } from './types';
 import { LocalTrackOptions } from './LocalTrackOptions';
 import { LocalTrackPublication } from './LocalTrackPublication';
 import { LocalVideoTrack } from './LocalVideoTrack';
@@ -11,26 +9,22 @@ import { Participant } from './Participant';
 import { Track } from './Track';
 import { TwilioError } from './TwilioError';
 
-export class LocalParticipant extends Participant {
+export class LocalParticipant extends Participant<LocalAudioTrackPublication, LocalDataTrackPublication, LocalVideoTrackPublication, LocalTrackPublication> {
   audioTracks: Map<Track.SID, LocalAudioTrackPublication>;
   dataTracks: Map<Track.SID, LocalDataTrackPublication>;
   tracks: Map<Track.SID, LocalTrackPublication>;
   videoTracks: Map<Track.SID, LocalVideoTrackPublication>;
   signalingRegion: string;
 
-  publishTrack(track: LocalTrack): Promise<LocalTrackPublication>;
-  publishTrack(mediaStreamTrack: MediaStreamTrack, options?: LocalTrackOptions): Promise<LocalTrackPublication>;
+  publishTrack(track: LocalTrack | MediaStreamTrack, options?: LocalTrackOptions): Promise<LocalTrackPublication>;
   publishTracks(tracks: Array<LocalTrack | MediaStreamTrack>): Promise<LocalTrackPublication[]>;
-  setNetworkQualityConfiguration(networkQualityConfiguration: NetworkQualityConfiguration): LocalParticipant;
-  setParameters(encodingParameters?: EncodingParameters | null): LocalParticipant;
-  unpublishTrack(track: LocalTrack | MediaStreamTrack): LocalTrackPublication;
+  setNetworkQualityConfiguration(networkQualityConfiguration: NetworkQualityConfiguration): this;
+  setParameters(encodingParameters?: EncodingParameters | null): this;
+  unpublishTrack(track: LocalTrack | MediaStreamTrack): LocalTrackPublication | null;
   unpublishTracks(tracks: Array<LocalTrack | MediaStreamTrack>): LocalTrackPublication[];
 
-  on(event: 'disconnected', listener: (participant: Participant) => void): this;
-  on(event: 'reconnected', listener: (participant: Participant) => void): this;
-  on(event: 'reconnecting', listener: (participant: Participant) => void): this;
+  on(event: 'disconnected', listener: (participant: this) => void): this;
   on(event: 'trackDimensionsChanged', listener: (track: LocalVideoTrack) => void): this;
-  on(event: 'networkQualityLevelChanged', listener: (networkQualityLevel: NetworkQualityLevel, networkQualityStats: NetworkQualityStats) => void): this;
   on(event: 'trackEnabled', listener: (track: LocalTrack) => void): this;
   on(event: 'trackDisabled', listener: (track: LocalTrack) => void): this;
   on(event: 'trackPublicationFailed', listener: (error: TwilioError, track: LocalTrack) => void): this;
