+++ tsdef/Participant.d.ts
@@ -1,5 +1,4 @@
-import { AudioTrackPublication, DataTrackPublication, VideoTrackPublication } from './types';
-import { NetworkQualityLevel, NetworkQualityStats } from './NetworkQuality';
+import { AudioTrackPublication, DataTrackPublication, NetworkQualityLevel, NetworkQualityStats, VideoTrackPublication } from './types';
 import { EventEmitter } from 'events';
 import { Track } from './Track';
 import { TrackPublication } from './TrackPublication';
@@ -10,7 +9,7 @@ export namespace Participant {
   type SID = string;
 }
 
-export class Participant extends EventEmitter {
+export class Participant<AudioTrackPublication, DataTrackPublication, VideoTrackPublication, TrackPublication> extends EventEmitter {
   audioTracks: Map<Track.SID, AudioTrackPublication>;
   dataTracks: Map<Track.SID, DataTrackPublication>;
   identity: Participant.Identity;
@@ -21,11 +20,11 @@ export class Participant extends EventEmitter {
   tracks: Map<Track.SID, TrackPublication>;
   videoTracks: Map<Track.SID, VideoTrackPublication>;
 
-  on(event: string, listener: (...args: any[]) => void): this;
-  on(event: 'disconnected', listener: (participant: Participant) => void): this;
-  on(event: 'reconnected', listener: (participant: Participant) => void): this;
-  on(event: 'reconnecting', listener: (participant: Participant) => void): this;
+  on(event: 'disconnected', listener: (participant: this) => void): this;
+  on(event: 'networkQualityLevelChanged', listener: (networkQualityLevel: NetworkQualityLevel, networkQualityStats: NetworkQualityStats) => void): this;
+  on(event: 'reconnected', listener: (participant: this) => void): this;
+  on(event: 'reconnecting', listener: (participant: this) => void): this;
   on(event: 'trackStarted', listener: (track: Track) => void): this;
   on(event: 'trackDimensionsChanged', listener: (track: VideoTrack) => void): this;
-  on(event: 'networkQualityLevelChanged', listener: (networkQualityLevel: NetworkQualityLevel, networkQualityStats: NetworkQualityStats) => void): this;
+  on(event: string, listener: (...args: any[]) => void): this;
 }
